Add Plasma wallpapers
Summary:
Add some of the best-looking Plasma wallpapers to this package so that users and
distros that install this package will have easier access to them. Especially for
distros that install this package by default, the result will be an aesthetic
enhancement.

The idea here is that when the `Next` wallpaper in the Breeze repo is updated with
a new one, the old one will be added here. That way this package will accumulate
the nicest Plasma wallpapers over time.

Test Plan:
Installed and saw that the new images were included with the others in the
`plasma-workspace-wallpapers` package.

Reviewers: #plasma, #vdg

Subscribers: abetts, yurchor, zzag, plasma-devel

Tags: #plasma

Maniphest Tasks: T9817

Differential Revision: https://phabricator.kde.org/D17780