Include only the largest size for each wallpaper
Summary:
In this patch, we remove all the versions of each wallpaper that are not the largest size. Here are the reasons why:
- **It's no longer necessary:** The old method appears to have been an artifact of the days before wallpapers could differ from the aspect ratio of the screen without looking bad. This is no longer a technical constraint that we face.
- **Simplicity:** It's easier to manage wallpapers, add new ones, and edit existing ones if you only have to do it for one image, not 15 or more.
- **Size savings:** installed size on disk falls from 152 MB to 25 MB.

There is no increase in memory use since Plasma automatically trims the image to the screen size before caching it in memory.

Test Plan:
All wallpapers are still installed and all still work and look good with the new "Scaled And Cropped" default setting.

Installed size is significantly smaller:
Before:
```
$ du -ch /usr/share/wallpapers/ | grep total
152M    total
```

After:
```
$ du -ch /usr/share/wallpapers/ | grep total
25M     total
```

Reviewers: #vdg, #plasma, ndavis

Reviewed By: #vdg, ndavis

Subscribers: GB_2, mart, rooty, sitter, rikmills, filipf, davidedmundson, cfeck, ndavis, plasma-devel

Tags: #plasma

Maniphest Tasks: T10269, T10219

Differential Revision: https://phabricator.kde.org/D18005