Add Dark Flow version from the original author of the light version
Interesting history tidbit: this was actually the original form of the
wallpaper that was subimtted, but at the time we didn't yet have
light/dark wallpaper changing capabilities, and people felt it was too
dark to use as a default wallpaper given that we use a light style by
default. Everyone agreed it would make a great dark-theme-specific
wallpaper though, so now that we have the ability to do those, let's use
it as one!


(cherry picked from commit 2ba678490bf7acd1c6ebc2ddccf733661210bc23)